requirements.txt
alembic==1.13.1
bitcoin==1.1.42
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
factory-boy==3.3.0
Faker==25.2.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.0.3
gunicorn==22.0.0
hupper==1.12.1
icdiff==2.0.7
idna==3.7
inflect==7.2.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.2.0
packaging==24.0
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
pluggy==1.5.0
pprintpp==0.4.0
psycopg2==2.9.9
pycoin==0.92.20230326
pycountry==23.12.11
pyramid==2.0.2
pyramid-basemodel==0.6.1
pyramid-mako==1.1.0
pyramid-tm==2.5
pyramid_localize==1.0.6
pytest==8.2.1
pytest-icdiff==0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.2
six==1.16.0
SQLAlchemy==2.0.30
text-unidecode==1.3
transaction==4.0
transactions==0.2.0
translationstring==1.4
typeguard==4.3.0
typing_extensions==4.12.0
urllib3==2.2.1
venusian==3.1.0
WebOb==1.8.7
Werkzeug==3.0.3
zope.deprecation==5.0
zope.interface==6.4.post2
zope.sqlalchemy==3.1
